+++ src/components/common/buttonContainer/index.js
@@ -1,0 +1,9 @@
+const ButtonContainer = (buttons) => {
+  const buttonContainer = document.createElement("div");
+  buttonContainer.classList.add("button-container");
+
+  buttons.forEach((button) => buttonContainer.appendChild(button));
+
+  return buttonContainer;
+};
+export default ButtonContainer;
